@@ -1,10 +1,17 @@
 module github.com/pangu-2/go-tools
 
-go 1.21
+go 1.24.1
 
 require (
 	github.com/gokits/gotools v1.0.0
-	github.com/matoous/go-nanoid/v2 v2.0.0
-	github.com/shopspring/decimal v1.3.1
+	github.com/matoous/go-nanoid/v2 v2.1.0
+	github.com/shopspring/decimal v1.4.0
 	github.com/thedevsaddam/gojsonq/v2 v2.5.2
+	gorm.io/gorm v1.30.0
+)
+
+require (
+	github.com/jinzhu/inflection v1.0.0 // indirect
+	github.com/jinzhu/now v1.1.5 // indirect
+	golang.org/x/text v0.20.0 // indirect
 )
